Real-time Sign Language Model
The objective of the Sign Recognition Project is to develop a deep-learning-based system
that accurately detects and interprets sign language gestures in real-time. The primary
goal is to bridge the communication gap between the deaf or hard-of-hearing community
and the wider society by providing a reliable and efficient means of translating sign
language into spoken or written language. The ultimate objective of sign language
recognition is to create inclusive environments where individuals who use sign language
can effectively communicate with others in various settings, such as educational
institutions, workplaces, healthcare facilities, or public spaces. The development of
robust and reliable sign language recognition systems can empower individuals with
hearing impairments, enhance their participation in society, and promote equal
opportunities for communication, education, and employment.